kafka_video_demo/kafka_python/requirements.txt
apturl==0.5.2
asttokens==2.2.1
backcall==0.2.0
blinker==1.4
blobfile==2.0.1
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
executing==1.2.0
filelock==3.9.0
fonttools==4.38.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.1
gyp==0.1
httplib2==0.20.2
idna==3.3
imageio==2.25.0
importlib-metadata==4.6.4
ipython==8.9.0
jedi==0.18.2
jeepney==0.7.1
kafka-python==2.0.2
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
more-itertools==8.10.0
netifaces==0.11.0
networkx==3.0
numpy==1.24.1
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.36
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycryptodomex==3.16.0
pycups==2.0.1
pydicom==2.3.1
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
PyWavelets==1.4.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
scikit-image==0.19.3
scipy==1.10.0
seaborn==0.12.2
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.46
ssh-import-id==5.11
stack-data==0.6.2
systemd-python==234
tifffile==2023.1.23.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.6
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
zope.interface==5.5.2
